requirements-dev.txt
aiokafka==0.10.0
cryptography==42.0.5
deprecated==1.2.14
email-validator==2.1.1
httpx==0.27.0
ipaddress==1.0.23
pydantic==2.7.1
pyyaml==6.0.1
setuptools==58.1.0

# Testing
pytest==8.2.0